Skips regenerating the index against template, stay on sketchy cogs
When you regen the pages via rake it also throws in the index and
births out this thing you don't actually see on the live site.

This will now just skip that and make things cleaner when rake-ing
and leave you with just those in the output p folder.

Fixes #29