Load and watch screencasts without leaving page
Adds a dirty amount of JavaScript and jQuery to catch the link
of the desired screencast and then swap out the embedded
YouTube video link. Also scrolls the page up in case you've
clicked an item that pushed the video out of full view.

If JavaScript is disabled on client-side this will still
allow the user to click through and view the video on YouTube.