Add link to YouTube channel that was replaced by screencasts page
The self-contained page of Git starter screencasts that brought
in the links from gitcasts.com replaced the raw link to YouTube
that was sitting on the homepage, so let's bring it back.

Also scopes the JavaScript to only pick out the links in the list
and not the rest of the div block. Otherwise you're never getting out.